requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
aiohttp==3.8.4
    # via openai
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.2
    # via aiohttp
attrs==23.1.0
    # via aiohttp
beautifulsoup4==4.12.2
    # via -r requirements.in
blinker==1.6.2
    # via flask
cachetools==5.3.1
    # via google-auth
certifi==2023.5.7
    # via
    #   httpx
    #   requests
chardet==3.0.4
    # via httpx
charset-normalizer==3.1.0
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via flask
flask==2.3.2
    # via -r requirements.in
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
google-api-core[grpc]==2.11.1
    # via
    #   google-cloud-core
    #   google-cloud-secret-manager
    #   google-cloud-storage
google-auth==2.20.0
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-core==2.3.2
    # via google-cloud-storage
google-cloud-secret-manager==2.16.2
    # via -r requirements.in
google-cloud-storage==2.9.0
    # via -r requirements.in
google-crc32c==1.5.0
    # via google-resumable-media
google-resumable-media==2.5.0
    # via google-cloud-storage
googleapis-common-protos[grpc]==1.59.1
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
googletrans==4.0.0rc1
    # via -r requirements.in
grpc-google-iam-v1==0.12.6
    # via google-cloud-secret-manager
grpcio==1.56.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.56.0
    # via google-api-core
h11==0.9.0
    # via httpcore
h2==3.2.0
    # via httpcore
hpack==3.0.0
    # via h2
hstspreload==2023.1.1
    # via httpx
httpcore==0.9.1
    # via httpx
httpx==0.13.3
    # via
    #   -r requirements.in
    #   googletrans
hyperframe==5.2.0
    # via h2
idna==2.10
    # via
    #   httpx
    #   requests
    #   yarl
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
openai==0.27.8
    # via -r requirements.in
proto-plus==1.22.3
    # via google-cloud-secret-manager
protobuf==4.23.3
    # via
    #   google-api-core
    #   google-cloud-secret-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
requests==2.31.0
    # via
    #   google-api-core
    #   google-cloud-storage
    #   openai
rfc3986==1.5.0
    # via httpx
rsa==4.9
    # via google-auth
six==1.16.0
    # via google-auth
sniffio==1.3.0
    # via
    #   httpcore
    #   httpx
soupsieve==2.4.1
    # via beautifulsoup4
tqdm==4.65.0
    # via openai
urllib3==1.26.16
    # via
    #   google-auth
    #   requests
werkzeug==2.3.6
    # via flask
yarl==1.9.2
    # via aiohttp
